Fix copy-to-clipboard
