+++ resources/assets/js/mix/copy_to_clipboard.js
@@ -7,3 +7,26 @@ function copyTextToClipboard(text) {
 }
 
 window.copyTextToClipboard = copyTextToClipboard;
+
+function fallbackCopyTextToClipboard(text) {
+  var textArea = document.createElement("textarea");
+  textArea.value = text;
+
+  // Avoid scrolling to bottom
+  textArea.style.top = "0";
+  textArea.style.left = "0";
+  textArea.style.position = "fixed";
+
+  document.body.appendChild(textArea);
+  textArea.focus();
+  textArea.select();
+
+  try {
+    document.execCommand("copy");
+  } catch (err) {
+    //Do nothing
+  }
+  document.body.removeChild(textArea);
+}
+
+window.fallbackCopyTextToClipboard = fallbackCopyTextToClipboard;
