+++ resources/assets/js/mix/filebrowser.js
@@ -1,4 +1,4 @@
-import { copyTextToClipboard } from "./copy_to_clipboard";
+require("./copy_to_clipboard.js");
 
 function itemReorder(id, pos, typ, act) {
   // Handles custom re-ordering of items in file browser
@@ -42,30 +42,9 @@ function itemReorder(id, pos, typ, act) {
 
 window.itemReorder = itemReorder;
 
-function fallbackCopyTextToClipboard(text) {
-  var textArea = document.createElement("textarea");
-  textArea.value = text;
-
-  // Avoid scrolling to bottom
-  textArea.style.top = "0";
-  textArea.style.left = "0";
-  textArea.style.position = "fixed";
-
-  document.body.appendChild(textArea);
-  textArea.focus();
-  textArea.select();
-
-  try {
-    document.execCommand("copy");
-  } catch (err) {
-    //Do nothing
-  }
-  document.body.removeChild(textArea);
-}
-
 function linkToClipboard(e) {
   var path = getSiteRoot() + e.dataset.title;
-  copyTextToClipboard(path);
+  window.copyTextToClipboard(path);
 }
 
 window.linkToClipboard = linkToClipboard;
